refactor: extract SalaryGrowthBadge and add to overpay page
